Finding a good address is more complicated than I realized.
There can be multiple street addresses returned with varying degrees
of accuracy.

There are also things called plus_codes which can be more useful than
``no address found.''

So score all of the values we get from a geocode request, then take
the best one.

It will take a while for addresses to update to the better values.