Rewrite how donuts are turned into bites.
* Make use of the spatial database for computations rather than pyproj/shapely
* Remove sprinkles being in overlapping bites by starting with COUNT sprinkles
  and reducing until size criteria is met

Not the speediest beast.  Even with the distance cache (which reduced the
calls to `ST_Distance()` by half), it still takes ~20 seconds to turn 4069
portals into 417 bites with a starting size of 20 portals/bite.

No profiling has been done to see what the hotspots are, but likely it is the
repeated calls back into spatialite.

Issue #14.