Start tracking "address types" as a ZCTA replacement.
This is collecting initial information for points to see how it looks across a
large sampling of portals.  The idea is that it would replace ZCTA as a way of
grouping portals.

In particular, the driving force behind this project was the nine Bay Area
counties.  This information includes the county names directly, which saves
having to manually look up new ZCTA codes and see if they should be kept or
not.