Create a new `PortalV2` table along with migrations.
This new table makes use of the newish SQL feature known as GENERATED or
COMPUTED columns.  Essentially, we are able to move the string parsing of
"lat,lng" pairs from Python into SQL.

The `latlng` column is a string again and the source of truth.  The removes
the standard issues surrounding dealing with floats and trying to match
against other tables.

The `point` column is used for spatial SQL work.  Testing seems to indicate
that everything was converted.

This includes migrating data from `Portal` to `PortalV2`.  No immediate plan
on retiring the old table.

Closes #21.