Enable full `mypy` processing for `database.py`.
This required an environmental configuration tweak.

It has long been known that the version of `typeshed` shipped with Debian
bookworm does not work well with SQLalchemy, with or without the mypy
extension.  And no good packaging existed that would address it.

However, if I `pip install sqlalchemy2-stubs`, then those stubs will get used
before those coming with the Debian version of `typeshed` and things magically
work.  I was quite surprised!

At this point, this the only package I install via `pip` (outside of this app
and `mundane`).  If it ever becomes an issue, I will extract and create a
local `.deb` for myself.