Run `PRAGMA optimize(-1)` on each connection close.
Now that we know a bit more, this can actually start to produce some useful
results.

First is that we now use `StaticPool` so values are valid for the whole
session.

Second is that `DELETE FROM from sqlite_stat1` will clear previous `ANALYZE`
values and so this will actually emit something while we tune it.

Issue #41.