Enhance the `portal show` flag `--bookmark` with features.
Previously this was a simple boolean flag.  Now it allows additional control
in how bookmark folders are handled.

The default is to fail if the folder already exists, keeping with the current
mode of operation.

However, it also possible for them to be cleared automatically or appendend
to (though the latter can still fail if duplicates are attempted).

Also, a user provided name can be used for simple bookmarks, rather than just
being stuck with `show`.

Closes #42.