Add "coverage" tests for easy stuff.
The word "coverage" is in quotes because these tests do not necessarily cover
all, or even much, of the logic implemented in the registration functions.

Eventually they should test that the right flags are generated, expected
constraints are present, and so on.  And some stuff, like `portal.py`'s
`QueryBuilder`, it really is not complete coverage.

But, like the last few changes, this is about getting line values up so
that *future* changes force the test coverage values up, presumably with real
tests.